Throw a DeepEqualException and move message building to its own class
